Separate generic web3 from veramo-specific functions
