+++ app/src/veramo-web3/Web3.tsx
@@ -9,54 +9,18 @@ import { UserRejectedRequestError as UserRejectedRequestErrorFrame } from '@web3
 import { Web3Provider } from '@ethersproject/providers'
 // import { formatEther } from '@ethersproject/units'
 
-import {
-  injected,
-  // network,
-  walletconnect,
-  // walletlink,
-  // ledger,
-  // trezor,
-  // lattice,
-  // frame,
-  // authereum,
-  // fortmatic,
-  // magic,
-  // portis,
-  // torus
-} from './connectors'
+import { injected, walletconnect } from '../wallets/connectors'
 import { Button } from 'antd'
 
 
 enum ConnectorNames {
   Injected = 'Metamask',
-  // Network = 'Network',
-  WalletConnect = 'WalletConnect',
-  // WalletLink = 'WalletLink',
-  // Ledger = 'Ledger',
-  // Trezor = 'Trezor',
-  // Lattice = 'Lattice',
-  // Frame = 'Frame',
-  // Authereum = 'Authereum',
-  // Fortmatic = 'Fortmatic',
-  // Magic = 'Magic',
-  // Portis = 'Portis',
-  // Torus = 'Torus'
+  WalletConnect = 'WalletConnect'
 }
 
 const connectorsByName: { [connectorName in ConnectorNames]: any } = {
   [ConnectorNames.Injected]: injected,
-  // [ConnectorNames.Network]: network,
-  [ConnectorNames.WalletConnect]: walletconnect,
-  // [ConnectorNames.WalletLink]: walletlink,
-  // [ConnectorNames.Ledger]: ledger,
-  // [ConnectorNames.Trezor]: trezor,
-  // [ConnectorNames.Lattice]: lattice,
-  // [ConnectorNames.Frame]: frame,
-  // [ConnectorNames.Authereum]: authereum,
-  // [ConnectorNames.Fortmatic]: fortmatic,
-  // [ConnectorNames.Magic]: magic,
-  // [ConnectorNames.Portis]: portis,
-  // [ConnectorNames.Torus]: torus
+  [ConnectorNames.WalletConnect]: walletconnect
 }
 
 function getErrorMessage(error: Error) {
@@ -76,158 +40,9 @@ function getErrorMessage(error: Error) {
   }
 }
 
-
-
-// function ChainId() {
-//   const { chainId } = useWeb3React()
-
-//   return (
-//     <>
-//       <span>Chain Id</span>
-//       <span role="img" aria-label="chain">
-//         ⛓
-//       </span>
-//       <span>{chainId ?? ''}</span>
-//     </>
-//   )
-// }
-
-// function BlockNumber() {
-//   const { chainId, library } = useWeb3React()
-
-//   const [blockNumber, setBlockNumber] = React.useState<number>()
-//   React.useEffect((): any => {
-//     if (!!library) {
-//       let stale = false
-
-//       library
-//         .getBlockNumber()
-//         .then((blockNumber: number) => {
-//           if (!stale) {
-//             setBlockNumber(blockNumber)
-//           }
-//         })
-//         .catch(() => {
-//           if (!stale) {
-//             setBlockNumber(undefined)
-//           }
-//         })
-
-//       const updateBlockNumber = (blockNumber: number) => {
-//         setBlockNumber(blockNumber)
-//       }
-//       library.on('block', updateBlockNumber)
-
-//       return () => {
-//         stale = true
-//         library.removeListener('block', updateBlockNumber)
-//         setBlockNumber(undefined)
-//       }
-//     }
-//   }, [library, chainId]) // ensures refresh if referential identity of library doesn't change across chainIds
-
-//   return (
-//     <>
-//       <span>Block Number</span>
-//       <span role="img" aria-label="numbers">
-//         🔢
-//       </span>
-//       <span>{blockNumber === null ? 'Error' : blockNumber ?? ''}</span>
-//     </>
-//   )
-// }
-
-// function Account() {
-//   const { account } = useWeb3React()
-
-//   return (
-//     <>
-//       <span>Account</span>
-//       <span role="img" aria-label="robot">
-//         🤖
-//       </span>
-//       <span>
-//         {account === null
-//           ? '-'
-//           : account
-//             ? `${account.substring(0, 6)}...${account.substring(account.length - 4)}`
-//             : ''}
-//       </span>
-//     </>
-//   )
-// }
-
-// function Balance() {
-//   const { account, library, chainId } = useWeb3React()
-
-//   const [balance, setBalance] = React.useState()
-//   React.useEffect((): any => {
-//     if (!!account && !!library) {
-//       let stale = false
-
-//       library
-//         .getBalance(account)
-//         .then((balance: any) => {
-//           if (!stale) {
-//             setBalance(balance)
-//           }
-//         })
-//         .catch(() => {
-//           if (!stale) {
-//             setBalance(undefined)
-//           }
-//         })
-
-//       return () => {
-//         stale = true
-//         setBalance(undefined)
-//       }
-//     }
-//   }, [account, library, chainId]) // ensures refresh if referential identity of library doesn't change across chainIds
-
-//   return (
-//     <>
-//       <span>Balance</span>
-//       <span role="img" aria-label="gold">
-//         💰
-//       </span>
-//       <span>{balance === null ? 'Error' : balance ? `Ξ${formatEther(balance as any)}` : ''}</span>
-//     </>
-//   )
-// }
-
-// function Header() {
-//   const { active, error } = useWeb3React()
-
-//   return (
-//     <>
-//       <h1 style={{ margin: '1rem', textAlign: 'right' }}>{active ? '🟢' : error ? '🔴' : '🟠'}</h1>
-//       <h3
-//         style={{
-//           display: 'grid',
-//           gridGap: '1rem',
-//           gridTemplateColumns: '1fr min-content 1fr',
-//           maxWidth: '20rem',
-//           lineHeight: '2rem',
-//           margin: 'auto'
-//         }}
-//       >
-//         <ChainId />
-//         <BlockNumber />
-//         <Account />
-//         <Balance />
-//       </h3>
-//     </>
-//   )
-// }
-
 export function Web3() {
   const context = useWeb3React<Web3Provider>()
-  const { connector,
-    // library,
-    // chainId, 
-    // account,
-     activate, deactivate, active, error } = context
+  const { connector, activate, deactivate, active, error } = context
 
   // handle logic to recognize the connector currently being activated
   const [activatingConnector, setActivatingConnector] = React.useState<any>()
